pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ultiplate"
version = "0.1.0"
description = "AI-native full-stack template with web, backend, worker, and ML apps"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "anthropic",
  "authlib>=1.6.9",
  "celery[redis]",
  "fastapi",
  "itsdangerous>=2.2.0",
  "opentelemetry-api",
  "opentelemetry-exporter-otlp-proto-http",
  "opentelemetry-instrumentation-fastapi",
  "opentelemetry-instrumentation-logging",
  "opentelemetry-instrumentation-requests",
  "opentelemetry-sdk",
  "psycopg[binary]>=3.3.3",
  "pydantic",
  "python-dotenv",
  "redis",
  "requests",
  "uvicorn[standard]",
]

[project.optional-dependencies]
flask = ["flask", "flask-cors"]
ml = ["tensorboard", "torch"]
mlflow = ["mlflow"]
scraping = ["beautifulsoup4", "seleniumbase"]
webapp-minimal = ["streamlit"]

[dependency-groups]
dev = [
  "black",
  "mypy",
  "pytest",
  "ruff",
]

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
include = ["alveslib*"]
exclude = ["apps*", "ml*", "tests*"]

[tool.uv]
default-groups = ["dev"]
